feat: built-in decimal → Decimal mapping via decimal.js (17a)
C# `decimal` now lowers to the `Decimal` class from the `decimal.js` package
by default, instead of silently degrading to the lossy JS `number` primitive.
The mapping ships in MetaSharp/Runtime/Decimal.cs as an [assembly: ExportFromBcl]
declaration; LoadBclExportMappings was extended to scan referenced assemblies
in addition to the current one, so the built-in flows transparently into every
consumer's compilation.

The user is responsible for adding `decimal.js` to their package.json — the
compiler only emits the import statement.

This commit handles the **type-level** mapping. Numeric literal lowering
(1.5m → new Decimal("1.5")) and operator lowering (a + b → a.plus(b) for
decimal operands) come in 17b/17c. Without those, code that does decimal
arithmetic will produce TypeScript that fails to compile, which is a fair
forcing function for the rest of the integration.

Two existing tests had to be updated:
- ExportFromBcl_WithoutMapping_UsesBuiltInFallback: the premise is now wrong
  (decimal has a built-in mapping). Renamed to Decimal_HasBuiltInMappingToDecimalJs
  and asserts the new default behavior.
- RecordWithMethods_GeneratesClassWithMethods: was using `decimal` incidentally;
  switched to `int` so it stays focused on record-class shape and doesn't
  exercise the new mapping in a way that would require literal/operator
  support that doesn't exist yet.

+++ MetaSharp.Tests/RecordTranspileTests.cs
@@ -17,13 +17,16 @@ public class RecordTranspileTests
     [Test]
     public async Task RecordWithMethods_GeneratesClassWithMethods()
     {
+        // Uses int (not decimal) so the assertion stays focused on record-class shape
+        // and doesn't accidentally exercise the decimal → Decimal mapping introduced
+        // for the decimal.js integration.
         var result = TranspileHelper.Transpile("""
             [Transpile]
-            public readonly record struct Amount(decimal Value)
+            public readonly record struct Amount(int Value)
             {
-                public decimal Doubled() => Value * 2;
+                public int Doubled() => Value * 2;
 
-                public static Amount FromValue(decimal value) => new(value);
+                public static Amount FromValue(int value) => new(value);
             }
             """);
 
@@ -36,9 +39,9 @@ public class RecordTranspileTests
     {
         var result = TranspileHelper.Transpile("""
             [Transpile]
-            public readonly record struct Amount(decimal Value)
+            public readonly record struct Amount(int Value)
             {
-                public decimal Doubled() => Value * 2;
+                public int Doubled() => Value * 2;
             }
             """);
